# docs/swagger.yaml
definitions:
  api.GenericResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /device/{udid}/accessibility/enable:
    get:
      description: Starts an accessibility session on the device and enables selection
        mode
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.GenericResponse'
      summary: Enable accessibility service
      tags:
      - accessibility
  /device/{udid}/accessibility/next:
    post:
      description: Moves the selection to the next element using accessibility service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.GenericResponse'
      summary: Navigate to next accessible element
      tags:
      - accessibility
  /device/{udid}/accessibility/perform-action:
    post:
      description: Performs the default action on the current focused element using
        DTX
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.GenericResponse'
      summary: Perform accessibility action via DTX
      tags:
      - accessibility
  /device/{udid}/accessibility/previous:
    post:
      description: Moves the selection to the previous element using accessibility
        service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.GenericResponse'
      summary: Navigate to previous accessible element
      tags:
      - accessibility
  /device/{udid}/accessibility/wda/perform-action:
    post:
      description: Performs an action using WebDriverAgent and returns action UUID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.GenericResponse'
      summary: Perform accessibility action via WDA
      tags:
      - accessibility
...
swagger: "2.0"